requirements.txt
ultralytics
streamlit
roboflow
opencv-python
numpy
pillow
matplotlib
